requirements.txt
Django==1.8.5
Pillow
OpenEXR